Update README.md to encourage people to use `Use this template`  (#1167)
This PR changes README.md to encourage people to use `Use this template`
rather than `fork`.

### Differences between `Use this template` and fork
> - A new fork includes the entire commit history of the parent
repository, while a repository created from a template starts with a
single commit.
> - Commits to a fork don't appear in your contributions graph, while
commits to a repository created from a template do appear in your
contribution graph.
> - A fork can be a temporary way to contribute code to an existing
project, while creating a repository from a template starts a new
project quickly.
[Ref](https://docs.github.com/repositories/creating-and-managing-repositories/creating-a-repository-from-a-template)

### Cons
Possible con would be that it is harder to pull the new changes from
this repo to the users' repos. It does not seem like impossible to do so
but it will require additonal actions on the users end.
- https://github.com/orgs/community/discussions/23528
-
https://stackoverflow.com/questions/56577184/github-pull-changes-from-a-template-repository

---------

Co-authored-by: Maruan <alshedivat@users.noreply.github.com>