update readme + remove sponsorship info (#1274)
majority of the recent improvements to al-folio have been truly driven
by a growing community of fans and contributors. this template has
gradually transformed from a low-key weekend project into a "public
good" of some sort used by thousands of people, university labs and
classes, and academic conference and workshops. as such, having personal
sponsorship info from the repository is no longer relevant.

as the community grows a bit further, we may have to think about
establishing some kind of governance around decisions of how al-folio is
supposed to evolve, which features to prioritize, and how to reward
contributors for their hard work. for now, these decisions will be made
by the project maintainers (active contributors are welcome to join the
team). and I'm all ears for proposals and suggestions!